backend/requirements.txt
# GigGatek Backend Dependencies

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Database
mysql-connector-python==8.1.0
protobuf==4.21.12

# Authentication
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
bcrypt==4.0.1

# CORS
Flask-CORS==4.0.0

# Development and Utilities
python-dotenv==1.0.0

# Push Notifications
pywebpush==1.14.0
py-vapid==1.9.0
cryptography==41.0.3
